requirements.txt
spotipy
requests
python-dotenv
psycopg2
alembic
tqdm